apps/workflow-engine/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
httpx
simpleeval
sqlmodel
sse-starlette
aiosqlite
pandas
openpyxl
google-genai
google-api-core
openai
anthropic
# aiosmtplib  # Uncomment if using SMTP for email
